+++ tests/Unit/AutoCreate/ProductAutoCreatorTest.php
@@ -27,17 +27,15 @@ final class ProductAutoCreatorTest extends TestCase {
 		WP_Mock::userFunction( 'get_option' )->andReturn(
 			array(
 				array(
-					'code'             => 'dmm-books',
-					'name'             => 'DMM Books',
-					'provider'         => 'dmm-ebook',
-					'displayOrder'     => 1,
-					'enabled'          => true,
-					'applicableTypes'  => array( 'ebook' ),
-					'buttonLabel'      => '',
-					'brandColor'       => '',
-					'buttonTextColor'  => '',
-					'autoRefresh'      => true,
-					'refreshFrequency' => 'weekly',
+					'code'            => 'dmm-books',
+					'name'            => 'DMM Books',
+					'provider'        => 'dmm-ebook',
+					'displayOrder'    => 1,
+					'enabled'         => true,
+					'applicableTypes' => array( 'ebook' ),
+					'buttonLabel'     => '',
+					'brandColor'      => '',
+					'buttonTextColor' => '',
 				),
 			)
 		);
@@ -80,6 +78,8 @@ final class ProductAutoCreatorTest extends TestCase {
 				$this->assertSame( '600', $data['listings'][0]['price'] );
 				$this->assertSame( 'auto', $data['listings'][0]['update_mode'] );
 				$this->assertTrue( $data['listings'][0]['auto_update'] );
+				$this->assertNotEmpty( $data['listings'][0]['last_verified_at'] );
+				$this->assertMatchesRegularExpression( '/^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}[+-]\d{2}:\d{2}$/', $data['listings'][0]['last_verified_at'] );
 				return 123;
 			}
 		);
@@ -91,17 +91,15 @@ final class ProductAutoCreatorTest extends TestCase {
 		WP_Mock::userFunction( 'get_option' )->andReturn(
 			array(
 				array(
-					'code'             => 'manual-shop',
-					'name'             => 'Manual',
-					'provider'         => 'manual',
-					'displayOrder'     => 1,
-					'enabled'          => true,
-					'applicableTypes'  => array( 'generic' ),
-					'buttonLabel'      => '',
-					'brandColor'       => '',
-					'buttonTextColor'  => '',
-					'autoRefresh'      => false,
-					'refreshFrequency' => 'weekly',
+					'code'            => 'manual-shop',
+					'name'            => 'Manual',
+					'provider'        => 'manual',
+					'displayOrder'    => 1,
+					'enabled'         => true,
+					'applicableTypes' => array( 'generic' ),
+					'buttonLabel'     => '',
+					'brandColor'      => '',
+					'buttonTextColor' => '',
 				),
 			)
 		);